+++ HISTORY.rst
@@ -2,7 +2,15 @@ Changelog
 =========
 
 
-1.0.0 (2023-01-??)
+1.1.0 (2023-04-04)
+~~~~~~~~~~~~~~~~~~
+
+* Split JS files into separate domain (like django does by default)
+
+* Support elm files. Should enable easy support of any ML-style language.
+
+
+1.0.0 (2023-02-17)
 ~~~~~~~~~~~~~~~~~~
 
 - Initial release